fix: onboarding loading state, session ID passing, reactive timetable day selection

+++ app/src/main/java/com/aditya/present/ui/screens/MainScreen.kt
@@ -46,7 +46,7 @@ import kotlinx.coroutines.launch
 
 @Composable
 fun MainScreen(
-    onAddSubject: () -> Unit,
+    onAddSubject: (Long) -> Unit,
     onAbout: () -> Unit = {},
 ) {
     val tabs = Tab.entries
